fix: update import paths for Breadcrumb, Pagination, and Spinner components; refactor review form handling
