+++ src/lib/prisma.ts
@@ -1,9 +1,5 @@
 import { PrismaClient } from '@prisma/client'
 
-// グローバル変数にPrismaクライアントを格納するための型定義
 const globalForPrisma = global as unknown as { prisma: PrismaClient }
-// Prismaクライアントインスタンスを作成または再利用
 export const prisma = globalForPrisma.prisma || new PrismaClient()
-
-// 開発環境では、グローバル変数にPrismaクライアントインスタンスを設定
 if (process.env.NODE_ENV !== 'production') globalForPrisma.prisma = prisma
